requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bravado==10.4.1
bravado-core==5.17.1
Bybit==0.2.12
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==39.0.2
dateparser==1.1.7
decorator==5.1.1
executing==1.2.0
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
ipython==8.11.0
isoduration==20.11.0
jedi==0.18.2
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
matplotlib-inline==0.1.6
monotonic==1.6
msgpack==1.0.5
msgpack-python==0.5.6
multidict==6.0.4
numpy==1.24.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pybit==2.4.1
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
pyrsistent==0.19.3
python-binance==1.0.17
python-dateutil==2.8.2
python-telegram-bot==12.8
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3987==1.3.8
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
swagger-spec-validator==3.0.3
TA-Lib==0.4.25
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
ujson==5.7.0
uri-template==1.2.0
urllib3==1.24.2
wcwidth==0.2.6
webcolors==1.12
websocket-client==1.5.1
websockets==10.4
yarl==1.8.2
